Move slashCommands startup module
